requirements.txt
streamlit
pandas
ezdxf
fpdf2
